requirements.txt
Pillow==10.2.0
python-magic==0.4.27
pytest
pytest-selenium